+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Previous version: 0.0.0b1
     - AddAtom operator class was added that adheres to this interface
     - RemoveAtom operator class was added that adheres to this interface
     - AddBond operator class was added that adheres to this interface
+    - RemoveBond operator class was added that adheres to this interface
 
 ## Fixes
 - number of linked dependencies was reduced (unused libraries are not linked now)